Refactored intel module
