+++ conan/tools/cmake/toolchain.py
@@ -10,6 +10,7 @@ from conan.tools._check_build_profile import check_using_build_profile
 from conan.tools._compilers import architecture_flag, use_win_mingw
 from conan.tools.cmake.utils import is_multi_configuration, get_file_name
 from conan.tools.files import save_toolchain_args
+from conan.tools.intel import IntelOneAPI
 from conan.tools.microsoft import VCVars
 from conan.tools.microsoft.visual import vs_ide_version
 from conans.errors import ConanException
@@ -530,6 +531,8 @@ class GenericSystemBlock(Block):
                 compiler_version = compiler_version if "." in compiler_version else \
                     "%s.0" % compiler_version
                 return "Intel C++ Compiler " + compiler_version
+        elif compiler == "intel-cc":
+            return IntelOneAPI(self._conanfile).ms_toolset
         elif compiler == "msvc":
             compiler_version = str(settings.compiler.version)
             version_components = compiler_version.split(".")
@@ -769,6 +772,9 @@ class CMakeToolchain(object):
         # Generators like Ninja or NMake requires an active vcvars
         if self.generator is not None and "Visual" not in self.generator:
             VCVars(self._conanfile).generate()
+        # If we're using Intel oneAPI, we need to generate the environment file and run it
+        elif self._conanfile.settings.get_safe("compiler") == "intel-cc":
+            IntelOneAPI(self._conanfile).generate()
         self._writebuild(toolchain_file)
 
     def _writebuild(self, toolchain_file):
