+++ conans/client/migrations_settings.py
@@ -2417,8 +2417,8 @@ compiler:
     intel-cc:
         version: ["2021.1", "2021.2", "2021.3"]
             update: [None, ANY]
-        mode: [None, "classic", "dpcpp"]
-        libcxx: [None, libstdc++, libstdc++11, libc++, c++_shared, c++_static]
+        mode: ["icx", "classic", "dpcpp"]
+        libcxx: [None, libstdc++, libstdc++11, libc++]
         cppstd: [None, 98, gnu98, 03, gnu03, 11, gnu11, 14, gnu14, 17, gnu17, 20, gnu20, 23, gnu23]
     qcc:
         version: ["4.4", "5.4", "8.3"]
